pyproject.toml
[tool.poetry]
name = "klusteri-website"
version = "0.1.0"
description = ""
authors = ["Laura Peltonen <laura.ms.peltonen@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.4"
coverage = "^7.4.0"
django = "^5.0.1"
pymongo = "^4.6.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
dnspython = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
